split SAML jaxrs config FAT class to avoid reconfig timeouts
